requirements.txt
# Core runtime dependencies — installed on every KACE machine
questionary==2.0.1
Jinja2==3.1.3
PyYAML==6.0.1

# SSH deployment (paramiko) is installed on-demand when the user
# first selects the SSH deploy option. See core/deployer.py.
# To pre-install SSH support manually: pip3 install paramiko==3.4.0
